third_party/WebKit/LayoutTests/inspector/console/console-uncaught-exception-expected.txt

CONSOLE ERROR: line 18: Uncaught Error: Exception in inline script.
CONSOLE ERROR: line 11: Uncaught Error: Exception in 'load' event listener.
CONSOLE ERROR: line 6: Uncaught Error: Exception in setTimeout callback.
Tests that uncaught exceptions are logged into console.Bug 47250.


uncaught-in-iframe.html:18 Uncaught Error: Exception in inline script.uncaught-in-iframe.html:18 auncaught-in-iframe.html:23 buncaught-in-iframe.html:26 (anonymous function)
uncaught-in-iframe.html:11 Uncaught Error: Exception in 'load' event listener.uncaught-in-iframe.html:11 funcaught-in-iframe.html:13 (anonymous function)
uncaught-in-iframe.html:6 Uncaught Error: Exception in setTimeout callback.uncaught-in-iframe.html:6 baruncaught-in-iframe.html:8 (anonymous function)
